Update sea.py
Adding in a scaling factor, k, for the population (in shortest_paths) for the first algorithm we need to implement. This makes it so that the population of the neighboring island has roughly a 50% influence (the other 50% coming from the distance to the island) on whether or not we choose that path.